Refactor main loop with undo manager
