+++ Undo.md
@@ -1,43 +1,28 @@
-# Undo/Redo Snapshot Framework
+# Undo and Redo
 
-This document describes the approach used to record user input for undo and redo
-actions in `ee`.
+`ee` keeps an in-memory history of snapshots.  Each snapshot stores the
+full text buffer and cursor state so a whole input chunk can be undone
+at once.  The history lives only for the current session.
 
-## Snapshotting
+## How it Works
 
-- A **snapshot** captures the entire buffer state along with cursor and screen
-  position information.
-- Snapshots are taken at the start of any input event that modifies text. This
-  includes character insertion, deletion, line operations and paste events.
-- Navigation commands such as arrow keys do not generate snapshots.
+1. `run_editor()` collects key presses in short bursts.  When more than
+   500 ms pass between keys or a paste is detected, the previous chunk is
+   closed with `undo_end_chunk()`.
+2. Editing functions call `start_action()` before modifying the buffer.
+   This uses `undo_begin_chunk()` to save a snapshot of the current
+   state the first time a chunk is touched.
+3. `undo_action()` and `redo_action()` swap snapshots between the undo
+   and redo stacks.
 
-## Input Granularity
+Large pastes are received as one chunk and therefore revert with a
+single undo.  Regular typing forms its own chunks so characters undo in
+order.
 
-Each physical input is grouped into an *undo chunk*:
+## Commands
 
-- Consecutive key presses within 500ms of each other extend the current chunk.
-- A paste operation is detected by reading the terminal buffer and always forms
-  a single chunk regardless of length or newlines.
+- **Ctrl+K** — undo the last chunk.
+- **Ctrl+R** — redo a previously undone chunk.
 
-Newlines are handled as ordinary text insertion so multi‑line pastes undo in a
-single step. This keeps the undo behaviour consistent when an entire block is
-pasted at once.
-
-## Stack Behaviour
-
-- `push_undo_state()` saves a snapshot to the undo stack and clears the redo
-  stack.
-- `undo_action()` restores the most recent snapshot and pushes the current state
-  onto the redo stack.
-- `redo_action()` restores the top snapshot from the redo stack and saves the
-  current state back to the undo stack.
-
-## Implementation Notes
-
-During input handling `collect_input_chunk()` gathers pending characters and
-waits briefly (30 ms) for more to arrive. This ensures large paste operations
-arrive as one array before processing begins. A timestamp check resets
-`last_action` and clears the chunk flag when more than 500 ms have elapsed since
-the previous input, starting a new undo group. `start_action()` takes a snapshot
-the first time it is called within a chunk so the whole block can be undone with
-one command.
+The stack depth is fixed at compile time (100 snapshots).  Older entries
+are dropped when the limit is exceeded.
